# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BloodConnect API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
paths:
  /auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
                full_name: { type: string }
                role: { type: string }
      responses:
        '200': { description: OK }
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: OK }
  /auth/refresh:
    post:
      summary: Refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        '200': { description: OK }
  /requests:
    get:
      summary: List my requests
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
    post:
      summary: Create a request
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                blood_group: { type: string }
                urgency: { type: string }
                notes: { type: string }
      responses:
        '201': { description: Created }
  /requests/{id}:
    put:
      summary: Update a request
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string }
                notes: { type: string }
      responses:
        '200': { description: OK }
    delete:
      summary: Delete a request
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          schema: { type: integer }
      responses:
        '200': { description: OK }
  /notifications/urgent:
    get:
      summary: Get urgent notifications
      responses:
        '200': { description: OK }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT